+++ examples/esp_zigbee_touchlink/touchlink_switch/main/esp_touchlink_switch.c
@@ -135,7 +135,7 @@ void esp_zb_app_signal_handler(esp_zb_app_signal_t *signal_struct)
         if (esp_zb_bdb_is_factory_new()) {
             esp_zb_start_touchlink_commissioning();
         } else {
-            ESP_LOGI(TAG, "Device restarted");
+            ESP_LOGI(TAG, "Device rebooted");
         }
         break;
     case ESP_ZB_BDB_SIGNAL_TOUCHLINK_NWK_STARTED: